Update documentation for traffic model
Also, added three more "total" counters, and removed the space again
in between the state name and the direction code ('+' or '-')

+++ test/test_traffic_model.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-import json
+import os, json
 from privcount.util import TrafficModel
 
 MODEL_FILENAME="traffic.model.json"
@@ -20,11 +20,15 @@ model = {
 }
 
 # write an uncompressed json file
-with open(MODEL_FILENAME, 'w') as outf:
-    json.dump(model, outf, sort_keys=True, separators=(',', ': '), indent=2)
+if not os.path.exists(MODEL_FILENAME):
+    with open(MODEL_FILENAME, 'w') as outf:
+        json.dump(model, outf, sort_keys=True, separators=(',', ': '), indent=2)
 
 del(model)
 model = None
+
+print "Testing traffic model..."
+print ""
 
 # now test reading in a model
 inf = open(MODEL_FILENAME, 'r')
@@ -39,10 +43,14 @@ emit_p = model['emission_probability']
 
 tmod = TrafficModel(states, start_p, trans_p, emit_p)
 
+print "Here is the list of all counter labels:"
+for label in sorted(tmod.get_counter_labels()):
+    print label
+print ""
+
 # sample observations
 observations = [('+', 20), ('+', 10), ('+', 50), ('+', 1000)]
 
+print "The most likly path through the traffic model given the observations is:"
 print "->".join(tmod.run_viterbi(observations))
-
-for label in sorted(tmod.get_counter_labels()):
-    print label
+print ""
